add abstract base class for composer scripthandler..
